Refactor provider component
